+++ provider/datasources/objectgroup.go
@@ -17,7 +17,7 @@ func DataSourceObjectGroup() *schema.Resource {
 		Schema: map[string]*schema.Schema{
 			"bucket": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Computed: true,
 			},
 			"public": {
 				Type:     schema.TypeBool,
@@ -77,6 +77,10 @@ func DataSourceObjectGroup() *schema.Resource {
 							Computed: true,
 						},
 						"field_selection": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"vertical_selection": {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
@@ -176,11 +180,15 @@ func DataSourceObjectGroup() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"ignore_irregular": {
-							Type:     schema.TypeBool,
+						"col_types": {
+							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"col_types": {
+						"col_renames": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"col_selection": {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
@@ -208,7 +216,7 @@ func DataSourceObjectGroup() *schema.Resource {
 			},
 			"partition_by": {
 				Type:     schema.TypeString,
-				Optional: true,
+				Computed: true,
 			},
 		},
 	}